+++ coders/cmyk.c
@@ -189,7 +189,7 @@ static Image *ReadCMYKImage(const ImageInfo *image_info,
               MagickMonitor(LoadImageText,y,image->rows);
         }
         count=image->tile_info.height-image->rows-image->tile_info.y;
-        for (i=0; i < count; i++)
+        for (i=0; i < (long) count; i++)
           (void) ReadBlob(image,packet_size*image->tile_info.width,scanline);
         break;
       }
@@ -227,7 +227,7 @@ static Image *ReadCMYKImage(const ImageInfo *image_info,
               MagickMonitor(LoadImageText,y,image->rows);
         }
         count=image->tile_info.height-image->rows-image->tile_info.y;
-        for (i=0; i < count; i++)
+        for (i=0; i < (long) count; i++)
           (void) ReadBlob(image,packet_size*image->tile_info.width,scanline);
         break;
       }
@@ -267,7 +267,7 @@ static Image *ReadCMYKImage(const ImageInfo *image_info,
           i++;
         }
         count=image->tile_info.height-image->rows-image->tile_info.y;
-        for (i=0; i < count; i++)
+        for (i=0; i < (long) count; i++)
           (void) ReadBlob(image,packet_size*image->tile_info.width,scanline);
         if (image_info->interlace == PartitionInterlace)
           {
@@ -293,7 +293,7 @@ static Image *ReadCMYKImage(const ImageInfo *image_info,
           i++;
         }
         count=image->tile_info.height-image->rows-image->tile_info.y;
-        for (i=0; i < count; i++)
+        for (i=0; i < (long) count; i++)
           (void) ReadBlob(image,packet_size*image->tile_info.width,scanline);
         if (image_info->interlace == PartitionInterlace)
           {
@@ -319,7 +319,7 @@ static Image *ReadCMYKImage(const ImageInfo *image_info,
           i++;
         }
         count=image->tile_info.height-image->rows-image->tile_info.y;
-        for (i=0; i < count; i++)
+        for (i=0; i < (long) count; i++)
           (void) ReadBlob(image,packet_size*image->tile_info.width,scanline);
         if (image_info->interlace == PartitionInterlace)
           {
@@ -345,7 +345,7 @@ static Image *ReadCMYKImage(const ImageInfo *image_info,
           i++;
         }
         count=image->tile_info.height-image->rows-image->tile_info.y;
-        for (i=0; i < count; i++)
+        for (i=0; i < (long) count; i++)
           (void) ReadBlob(image,packet_size*image->tile_info.width,scanline);
         if (image->matte)
           {
@@ -379,7 +379,7 @@ static Image *ReadCMYKImage(const ImageInfo *image_info,
               i++;
             }
             count=image->tile_info.height-image->rows-image->tile_info.y;
-            for (i=0; i < count; i++)
+            for (i=0; i < (long) count; i++)
               (void) ReadBlob(image,packet_size*image->tile_info.width,
                 scanline);
           }
