+++ coders/miff.c
@@ -698,7 +698,7 @@ static Image *ReadMIFFImage(const ImageInfo *image_info,
         /*
           Generic profile.
         */
-        for (i=0; i < image->generic_profiles; i++)
+        for (i=0; i < (long) image->generic_profiles; i++)
         {
           if (image->generic_profile[i].length == 0)
             continue;
@@ -739,14 +739,14 @@ static Image *ReadMIFFImage(const ImageInfo *image_info,
             (void) ReadBlob(image,packet_size*image->colors,colormap);
             p=colormap;
             if (image->depth <= 8)
-              for (i=0; i < image->colors; i++)
+              for (i=0; i < (long) image->colors; i++)
               {
                 image->colormap[i].red=UpScale(*p++);
                 image->colormap[i].green=UpScale(*p++);
                 image->colormap[i].blue=UpScale(*p++);
               }
             else
-              for (i=0; i < image->colors; i++)
+              for (i=0; i < (long) image->colors; i++)
               {
                 image->colormap[i].red=(*p++ << 8);
                 image->colormap[i].red|=(*p++);
@@ -1391,7 +1391,7 @@ static unsigned int WriteMIFFImage(const ImageInfo *image_info,Image *image)
         /*
           Generic profile.
         */
-        for (i=0; i < image->generic_profiles; i++)
+        for (i=0; i < (long) image->generic_profiles; i++)
         {
           FormatString(buffer,"profile-%.1024s=%u\n",
             image->generic_profile[i].name == (char *) NULL ? "generic" :
@@ -1412,13 +1412,13 @@ static unsigned int WriteMIFFImage(const ImageInfo *image_info,Image *image)
         continue;
       FormatString(buffer,"%.1024s=",attribute->key);
       (void) WriteBlobString(image,buffer);
-      for (i=0; i < strlen(attribute->value); i++)
+      for (i=0; i < (long) strlen(attribute->value); i++)
         if (isspace((int) attribute->value[i]))
           break;
-      if (i < strlen(attribute->value))
+      if (i < (long) strlen(attribute->value))
         (void) WriteBlobByte(image,'{');
       (void) WriteBlob(image,strlen(attribute->value),attribute->value);
-      if (i < strlen(attribute->value))
+      if (i < (long) strlen(attribute->value))
         (void) WriteBlobByte(image,'}');
       (void) WriteBlobByte(image,'\n');
     }
@@ -1443,7 +1443,7 @@ static unsigned int WriteMIFFImage(const ImageInfo *image_info,Image *image)
         /*
           Generic profile.
         */
-        for (i=0; i < image->generic_profiles; i++)
+        for (i=0; i < (long) image->generic_profiles; i++)
         {
           if (image->generic_profile[i].length == 0)
             continue;
@@ -1475,14 +1475,14 @@ static unsigned int WriteMIFFImage(const ImageInfo *image_info,Image *image)
         */
         q=colormap;
         if (image->depth <= 8)
-          for (i=0; i < image->colors; i++)
+          for (i=0; i < (long) image->colors; i++)
           {
             *q++=DownScale(image->colormap[i].red);
             *q++=DownScale(image->colormap[i].green);
             *q++=DownScale(image->colormap[i].blue);
           }
         else
-          for (i=0; i < image->colors; i++)
+          for (i=0; i < (long) image->colors; i++)
           {
             *q++=image->colormap[i].red >> 8;
             *q++=image->colormap[i].red & 0xff;
